refactor(build): rename build/ to vite-stubs/ to avoid gitignore conflict
The src/lib/build/ directory was being caught by the root-level
build/ gitignore pattern. Renamed to src/lib/vite-stubs/ which
better describes its purpose and avoids the conflict.
