+++ src/lib/vite-stubs/browser-stubs.ts
